+++ src/utils/jira/getMostRelevantJiraTickets.ts
@@ -18,22 +18,5 @@ export default async function getMostRelevantJiraTickets({
   if (jiraTickets?.error === "no access_token") {
     return { errorText: "Not logged in" };
   }
-  let ticketsWithComments: any[] = [];
-  if (!jiraTickets || jiraTickets.length === 0 || jiraTickets.error) {
-    return [];
-  }
-  let ticketPromises = jiraTickets?.map(
-    async (ticket: { key: string; comments: any[] }) => {
-      let comments = await getTicketComments({
-        email: user,
-        issueIdOrKey: ticket.key,
-      });
-      ticketsWithComments.push({
-        ...ticket,
-        comments,
-      });
-    }
-  );
-  await Promise.all(ticketPromises);
-  return ticketsWithComments;
+  return jiraTickets;
 }